Further deamalgamate src/sx
The `src/sx` folder previously was deamalgamated from `src/ph7/lib.c`,
which was deamalgamated from `ph7.c`.

Now, we further untangle the sx library into individual modules, each
own with individual interfaces.
