+++ src/sx/sxbase64.h
@@ -1,0 +1,17 @@
+/**
+ * SPDX-FileCopyrightText: 2011, 2012, 2013, 2014 Symisc Systems <licensing@symisc.net>
+ * SPDX-FileCopyrightText: 2025 Alexandre Gomes Gaigalas <alganet@gmail.com>
+ * SPDX-License-Identifier: BSD-3-Clause
+ */
+#ifndef __SXBASE64_H__
+#define __SXBASE64_H__
+
+#include "sxtypes.h"
+
+/* Base64 function prototypes */
+#ifndef PH7_DISABLE_BUILTIN_FUNC
+PH7_PRIVATE sxi32 SyBase64Encode(const char *zSrc,sxu32 nLen,ProcConsumer xConsumer,void *pUserData);
+PH7_PRIVATE sxi32 SyBase64Decode(const char *zB64,sxu32 nLen,ProcConsumer xConsumer,void *pUserData);
+#endif /* PH7_DISABLE_BUILTIN_FUNC */
+
+#endif /* __SXBASE64_H__ */
