+++ src/sx/sxds.c
@@ -3,7 +3,13 @@
  * SPDX-FileCopyrightText: 2025 Alexandre Gomes Gaigalas <alganet@gmail.com>
  * SPDX-License-Identifier: BSD-3-Clause
  */
-#include "sxproto.h"
+#include "sxtypes.h"
+#include "sxmacros.h"
+#include "sxset.h"
+#include "sxmem.h"
+#include "sxhashtable.h"
+#include "sxhash.h"
+#include "sxstr.h"
 
 PH7_PRIVATE sxi32 SySetInit(SySet *pSet,SyMemBackend *pAllocator,sxu32 ElemSize)
 {
