+++ src/sx/sxmem.c
@@ -3,7 +3,12 @@
  * SPDX-FileCopyrightText: 2025 Alexandre Gomes Gaigalas <alganet@gmail.com>
  * SPDX-License-Identifier: BSD-3-Clause
  */
-#include "sxproto.h"
+#include "sxtypes.h"
+#include "sxmacros.h"
+#include "sxset.h"
+#include "sxmem.h"
+#include "sxmutex.h"
+#include "sxstr.h"
 #if defined(__WINNT__)
 #include <Windows.h>
 #else
