--- src/sx/sxproto.h
@@ -1,167 +1,0 @@
-/**
- * SPDX-FileCopyrightText: 2011, 2012, 2013, 2014 Symisc Systems <licensing@symisc.net>
- * SPDX-FileCopyrightText: 2025 Alexandre Gomes Gaigalas <alganet@gmail.com>
- * SPDX-License-Identifier: BSD-3-Clause
- */
-#ifndef __SXPROTO_H__
-#define __SXPROTO_H__
-
-/* 
- * Symisc library function prototypes.
- * This header declares all Sy* prefixed functions from the Symisc library.
- */
-
-#include "sxint.h"
-
-/* Define PH7_PRIVATE for lib internal functions */
-#ifndef PH7_PRIVATE
-#define PH7_PRIVATE
-#endif
-
-/* lib.c function prototypes */
-#ifndef PH7_DISABLE_BUILTIN_FUNC
-PH7_PRIVATE sxi32 SyXMLParserInit(SyXMLParser *pParser,SyMemBackend *pAllocator,sxi32 iFlags);
-PH7_PRIVATE sxi32 SyXMLParserSetEventHandler(SyXMLParser *pParser,
-	void *pUserData,
-	ProcXMLStartTagHandler xStartTag,
-	ProcXMLTextHandler xRaw,
-	ProcXMLSyntaxErrorHandler xErr,
-	ProcXMLStartDocument xStartDoc,
-	ProcXMLEndTagHandler xEndTag,
-	ProcXMLPIHandler xPi,
-	ProcXMLEndDocument xEndDoc,
-	ProcXMLDoctypeHandler xDoctype,
-	ProcXMLNameSpaceStart xNameSpace,
-	ProcXMLNameSpaceEnd xNameSpaceEnd
-	);
-PH7_PRIVATE sxi32 SyXMLProcess(SyXMLParser *pParser,const char *zInput,sxu32 nByte);
-PH7_PRIVATE sxi32 SyXMLParserRelease(SyXMLParser *pParser);
-PH7_PRIVATE sxi32 SyArchiveInit(SyArchive *pArch,SyMemBackend *pAllocator,ProcHash xHash,ProcRawStrCmp xCmp);
-PH7_PRIVATE sxi32 SyArchiveRelease(SyArchive *pArch);
-PH7_PRIVATE sxi32 SyArchiveResetLoopCursor(SyArchive *pArch);
-PH7_PRIVATE sxi32 SyArchiveGetNextEntry(SyArchive *pArch,SyArchiveEntry **ppEntry);
-PH7_PRIVATE sxi32 SyZipExtractFromBuf(SyArchive *pArch,const char *zBuf,sxu32 nLen);
-#endif /* PH7_DISABLE_BUILTIN_FUNC */
-#ifndef PH7_DISABLE_BUILTIN_FUNC
-PH7_PRIVATE sxi32 SyBinToHexConsumer(const void *pIn,sxu32 nLen,ProcConsumer xConsumer,void *pConsumerData);
-#endif /* PH7_DISABLE_BUILTIN_FUNC */
-#ifndef PH7_DISABLE_BUILTIN_FUNC
-#ifndef PH7_DISABLE_HASH_FUNC
-PH7_PRIVATE sxu32 SyCrc32(const void *pSrc,sxu32 nLen);
-PH7_PRIVATE void MD5Update(MD5Context *ctx, const unsigned char *buf, unsigned int len);
-PH7_PRIVATE void MD5Final(unsigned char digest[16], MD5Context *ctx);
-PH7_PRIVATE sxi32 MD5Init(MD5Context *pCtx);
-PH7_PRIVATE sxi32 SyMD5Compute(const void *pIn,sxu32 nLen,unsigned char zDigest[16]);
-PH7_PRIVATE void SHA1Init(SHA1Context *context);
-PH7_PRIVATE void SHA1Update(SHA1Context *context,const unsigned char *data,unsigned int len);
-PH7_PRIVATE void SHA1Final(SHA1Context *context, unsigned char digest[20]);
-PH7_PRIVATE sxi32 SySha1Compute(const void *pIn,sxu32 nLen,unsigned char zDigest[20]);
-#endif
-#endif /* PH7_DISABLE_BUILTIN_FUNC */
-PH7_PRIVATE sxi32 SyRandomness(SyPRNGCtx *pCtx,void *pBuf,sxu32 nLen);
-PH7_PRIVATE sxi32 SyRandomnessInit(SyPRNGCtx *pCtx,ProcRandomSeed xSeed,void *pUserData);
-PH7_PRIVATE sxu32 SyBufferFormat(char *zBuf,sxu32 nLen,const char *zFormat,...);
-PH7_PRIVATE sxu32 SyBlobFormatAp(SyBlob *pBlob,const char *zFormat,va_list ap);
-PH7_PRIVATE sxu32 SyBlobFormat(SyBlob *pBlob,const char *zFormat,...);
-PH7_PRIVATE sxi32 SyProcFormat(ProcConsumer xConsumer,void *pData,const char *zFormat,...);
-#ifndef PH7_DISABLE_BUILTIN_FUNC
-PH7_PRIVATE const char *SyTimeGetMonth(sxi32 iMonth);
-PH7_PRIVATE const char *SyTimeGetDay(sxi32 iDay);
-#endif /* PH7_DISABLE_BUILTIN_FUNC */
-PH7_PRIVATE sxi32 SyUriDecode(const char *zSrc,sxu32 nLen,ProcConsumer xConsumer,void *pUserData,int bUTF8);
-#ifndef PH7_DISABLE_BUILTIN_FUNC
-PH7_PRIVATE sxi32 SyUriEncode(const char *zSrc,sxu32 nLen,ProcConsumer xConsumer,void *pUserData);
-#endif
-PH7_PRIVATE sxi32 SyLexRelease(SyLex *pLex);
-PH7_PRIVATE sxi32 SyLexTokenizeInput(SyLex *pLex,const char *zInput,sxu32 nLen,void *pCtxData,ProcSort xSort,ProcCmp xCmp);
-PH7_PRIVATE sxi32 SyLexInit(SyLex *pLex,SySet *pSet,ProcTokenizer xTokenizer,void *pUserData);
-#ifndef PH7_DISABLE_BUILTIN_FUNC
-PH7_PRIVATE sxi32 SyBase64Decode(const char *zB64,sxu32 nLen,ProcConsumer xConsumer,void *pUserData);
-PH7_PRIVATE sxi32 SyBase64Encode(const char *zSrc,sxu32 nLen,ProcConsumer xConsumer,void *pUserData);
-#endif /* PH7_DISABLE_BUILTIN_FUNC */
-PH7_PRIVATE sxi32 SyStrToReal(const char *zSrc,sxu32 nLen,void *pOutVal,const char **zRest);
-PH7_PRIVATE sxi32 SyBinaryStrToInt64(const char *zSrc,sxu32 nLen,void *pOutVal,const char **zRest);
-PH7_PRIVATE sxi32 SyOctalStrToInt64(const char *zSrc,sxu32 nLen,void *pOutVal,const char **zRest);
-PH7_PRIVATE sxi32 SyHexStrToInt64(const char *zSrc,sxu32 nLen,void *pOutVal,const char **zRest);
-PH7_PRIVATE sxi32 SyHexToint(sxi32 c);
-PH7_PRIVATE sxi32 SyStrToInt64(const char *zSrc,sxu32 nLen,void *pOutVal,const char **zRest);
-PH7_PRIVATE sxi32 SyStrToInt32(const char *zSrc,sxu32 nLen,void *pOutVal,const char **zRest);
-PH7_PRIVATE sxi32 SyStrIsNumeric(const char *zSrc,sxu32 nLen,sxu8 *pReal,const char **pzTail);
-PH7_PRIVATE SyHashEntry *SyHashLastEntry(SyHash *pHash);
-PH7_PRIVATE sxi32 SyHashInsert(SyHash *pHash,const void *pKey,sxu32 nKeyLen,void *pUserData);
-PH7_PRIVATE sxi32 SyHashForEach(SyHash *pHash,sxi32(*xStep)(SyHashEntry *,void *),void *pUserData);
-PH7_PRIVATE SyHashEntry *SyHashGetNextEntry(SyHash *pHash);
-PH7_PRIVATE sxi32 SyHashResetLoopCursor(SyHash *pHash);
-PH7_PRIVATE sxi32 SyHashDeleteEntry2(SyHashEntry *pEntry);
-PH7_PRIVATE sxi32 SyHashDeleteEntry(SyHash *pHash,const void *pKey,sxu32 nKeyLen,void **ppUserData);
-PH7_PRIVATE SyHashEntry *SyHashGet(SyHash *pHash,const void *pKey,sxu32 nKeyLen);
-PH7_PRIVATE sxi32 SyHashRelease(SyHash *pHash);
-PH7_PRIVATE sxi32 SyHashInit(SyHash *pHash,SyMemBackend *pAllocator,ProcHash xHash,ProcCmp xCmp);
-PH7_PRIVATE sxu32 SyStrHash(const void *pSrc,sxu32 nLen);
-PH7_PRIVATE sxu32 SyBinHash(const void *pSrc,sxu32 nLen);
-PH7_PRIVATE sxu32 Systrcpy(char *zDest,sxu32 nDestLen,const char *zSrc,sxu32 nLen);
-PH7_PRIVATE void *SySetAt(SySet *pSet,sxu32 nIdx);
-PH7_PRIVATE void *SySetPop(SySet *pSet);
-PH7_PRIVATE void *SySetPeek(SySet *pSet);
-PH7_PRIVATE sxi32 SySetRelease(SySet *pSet);
-PH7_PRIVATE sxi32 SySetReset(SySet *pSet);
-PH7_PRIVATE sxi32 SySetResetCursor(SySet *pSet);
-PH7_PRIVATE sxi32 SySetGetNextEntry(SySet *pSet,void **ppEntry);
-#ifndef PH7_DISABLE_BUILTIN_FUNC
-PH7_PRIVATE void * SySetPeekCurrentEntry(SySet *pSet);
-#endif /* PH7_DISABLE_BUILTIN_FUNC */
-PH7_PRIVATE sxi32 SySetTruncate(SySet *pSet,sxu32 nNewSize);
-PH7_PRIVATE sxi32 SySetAlloc(SySet *pSet,sxi32 nItem);
-PH7_PRIVATE sxi32 SySetPut(SySet *pSet,const void *pItem);
-PH7_PRIVATE sxi32 SySetInit(SySet *pSet,SyMemBackend *pAllocator,sxu32 ElemSize);
-#ifndef PH7_DISABLE_BUILTIN_FUNC
-PH7_PRIVATE sxi32 SyBlobSearch(const void *pBlob,sxu32 nLen,const void *pPattern,sxu32 pLen,sxu32 *pOfft);
-#endif
-PH7_PRIVATE sxi32 SyBlobRelease(SyBlob *pBlob);
-PH7_PRIVATE sxi32 SyBlobReset(SyBlob *pBlob);
-PH7_PRIVATE sxi32 SyBlobCmp(SyBlob *pLeft,SyBlob *pRight);
-PH7_PRIVATE sxi32 SyBlobDup(SyBlob *pSrc,SyBlob *pDest);
-PH7_PRIVATE sxi32 SyBlobNullAppend(SyBlob *pBlob);
-PH7_PRIVATE sxi32 SyBlobAppend(SyBlob *pBlob,const void *pData,sxu32 nSize);
-PH7_PRIVATE sxi32 SyBlobReadOnly(SyBlob *pBlob,const void *pData,sxu32 nByte);
-PH7_PRIVATE sxi32 SyBlobInit(SyBlob *pBlob,SyMemBackend *pAllocator);
-PH7_PRIVATE sxi32 SyBlobInitFromBuf(SyBlob *pBlob,void *pBuffer,sxu32 nSize);
-PH7_PRIVATE char *SyMemBackendStrDup(SyMemBackend *pBackend,const char *zSrc,sxu32 nSize);
-PH7_PRIVATE void *SyMemBackendDup(SyMemBackend *pBackend,const void *pSrc,sxu32 nSize);
-PH7_PRIVATE sxi32 SyMemBackendRelease(SyMemBackend *pBackend);
-PH7_PRIVATE sxi32 SyMemBackendInitFromOthers(SyMemBackend *pBackend,const SyMemMethods *pMethods,ProcMemError xMemErr,void *pUserData);
-PH7_PRIVATE sxi32 SyMemBackendInit(SyMemBackend *pBackend,ProcMemError xMemErr,void *pUserData);
-PH7_PRIVATE sxi32 SyMemBackendInitFromParent(SyMemBackend *pBackend,SyMemBackend *pParent);
-#if 0
-/* Not used in the current release of the PH7 engine */
-PH7_PRIVATE void *SyMemBackendPoolRealloc(SyMemBackend *pBackend,void *pOld,sxu32 nByte);
-#endif
-PH7_PRIVATE sxi32 SyMemBackendPoolFree(SyMemBackend *pBackend,void *pChunk);
-PH7_PRIVATE void *SyMemBackendPoolAlloc(SyMemBackend *pBackend,sxu32 nByte);
-PH7_PRIVATE sxi32 SyMemBackendFree(SyMemBackend *pBackend,void *pChunk);
-PH7_PRIVATE void *SyMemBackendRealloc(SyMemBackend *pBackend,void *pOld,sxu32 nByte);
-PH7_PRIVATE void *SyMemBackendAlloc(SyMemBackend *pBackend,sxu32 nByte);
-#if defined(PH7_ENABLE_THREADS)
-PH7_PRIVATE sxi32 SyMemBackendMakeThreadSafe(SyMemBackend *pBackend,const SyMutexMethods *pMethods);
-PH7_PRIVATE sxi32 SyMemBackendDisbaleMutexing(SyMemBackend *pBackend);
-#endif
-PH7_PRIVATE sxu32 SyMemcpy(const void *pSrc,void *pDest,sxu32 nLen);
-PH7_PRIVATE sxi32 SyMemcmp(const void *pB1,const void *pB2,sxu32 nSize);
-PH7_PRIVATE void SyZero(void *pSrc,sxu32 nSize);
-PH7_PRIVATE sxi32 SyStrnicmp(const char *zLeft,const char *zRight,sxu32 SLen);
-PH7_PRIVATE sxi32 SyStrnmicmp(const void *pLeft, const void *pRight,sxu32 SLen);
-#ifndef PH7_DISABLE_BUILTIN_FUNC
-PH7_PRIVATE sxi32 SyStrncmp(const char *zLeft,const char *zRight,sxu32 nLen);
-#endif
-PH7_PRIVATE sxi32 SyByteListFind(const char *zSrc,sxu32 nLen,const char *zList,sxu32 *pFirstPos);
-#ifndef PH7_DISABLE_BUILTIN_FUNC
-PH7_PRIVATE sxi32 SyByteFind2(const char *zStr,sxu32 nLen,sxi32 c,sxu32 *pPos);
-#endif
-PH7_PRIVATE sxi32 SyByteFind(const char *zStr,sxu32 nLen,sxi32 c,sxu32 *pPos);
-PH7_PRIVATE sxu32 SyStrlen(const char *zSrc);
-#if defined(PH7_ENABLE_THREADS)
-PH7_PRIVATE const SyMutexMethods *SyMutexExportMethods(void);
-PH7_PRIVATE sxi32 SyMemBackendMakeThreadSafe(SyMemBackend *pBackend,const SyMutexMethods *pMethods);
-PH7_PRIVATE sxi32 SyMemBackendDisbaleMutexing(SyMemBackend *pBackend);
-#endif
-#endif /* __SXPROTO_H__ */
