+++ src/sx/sxstr.c
@@ -3,7 +3,9 @@
  * SPDX-FileCopyrightText: 2025 Alexandre Gomes Gaigalas <alganet@gmail.com>
  * SPDX-License-Identifier: BSD-3-Clause
  */
-#include "sxproto.h"
+#include "sxtypes.h"
+#include "sxmacros.h"
+#include "sxstr.h"
 
 PH7_PRIVATE sxu32 SyStrlen(const char *zSrc)
 {
